Workarounded nimibook bug
Nimibook does not properly render code-block in nbText blocks.
It will render the ` characters in the text and treat every following
text as part of the code-block.

To avoid this issue, you can lcose the nbText block and open a new one.
To avoid the ` to be rendered, you can just leave them out.