Refactor placement of StyleClass into widgets
Was previously located in widgetutils.
However, other places use it as well that don't import widgetutils, but widgets.
It seems iffy to change that chain of dependencies.
Further, Margin is in a similar situation and is defined in widgets.nim
It seems more consistent to thusly do it there as well.