#82 Add custom pragma "locker" to disguise "Locks:0"
This is a measure for better support of both nim 1.6. and nim 2.0
2.0 no longer supports lock-levels where they are deprecated.
So the locks:0 pragma is also deprecated and should be removed.
Nim 1.6. meanwhile will spit out warnings over undefined lock levels
if the examples are compiled without the pragma.

Thus customPragmas was created to define a `locker` pragma.
That pragma does nothing on nim2.0 and contains the locks:0 pragma
on nim 1.6., satisfying both versions and limiting warning noise.