Update `=destroy` hooks and fix lock level warnings (#81)
* debug attempt

* #75 initial value based semantics version for new destructors

* #75 second attempt using pointer semantics

This is a version of using destructors where PixBuf
is not a value object, but the pointer to GdkPixbuf directly.
You can simply revert this commit to return to the prior version
where PixBuf is a value object wrapping GdkPixBuf.
That version is known to work.

* Add =sink hook and change GdkPixBuf to GdkPixbuf

* Revert "Add =sink hook and change GdkPixBuf to GdkPixbuf"

This reverts commit ecf8f6d550655082f4dd28d6c377e37611a3e29f.

* Revert "#75 second attempt using pointer semantics"

This reverts commit fbb70a9457f7f8767d101816a56880c6f3668154.

* #75 Third approach for destructors

We wrap the gdk pointer in a ref object instead of an object.

PixbufObj becomes an intermediate value-object and Pixbuf
an alias to the ref-object version of PixbufObj.

* #75 remove pointless docs.yml change

* #75 Make examples compile against project owlkettle version

By adding the path flag to the nim compilation command
we can ensure that the owlkettle version used to compile
the examples is the **local** version, not one globally installed.

* #82 Add custom pragma "locker" to disguise "Locks:0"

This is a measure for better support of both nim 1.6. and nim 2.0
2.0 no longer supports lock-levels where they are deprecated.
So the locks:0 pragma is also deprecated and should be removed.
Nim 1.6. meanwhile will spit out warnings over undefined lock levels
if the examples are compiled without the pragma.

Thus customPragmas was created to define a `locker` pragma.
That pragma does nothing on nim2.0 and contains the locks:0 pragma
on nim 1.6., satisfying both versions and limiting warning noise.

* Nim 2.0 desires its destructor hooks to be defined as
proc `=destroy`(x: T) =

Nim 1.6 desires its destructor hooks to be defined as
proc `=destroy`(x: var T) =

That is deprecated in nim 2.0.
To satisfy both, the new destructor hooks were put behind
a when statement

* #76 Add destructors for Stylesheet

* #82 Refactor how locker pragma gets generated and provided

* #82 Add doc comment

* #82 Remove now pointless file

* #75 Refactor generating multi-nim-version hooks via template

* #75 streamline construction of destructor objects

* #75 adjust indentation

* #75 add copy hook for TextBuffer

---------

Co-authored-by: Can Lehmann <can.l@posteo.de>