#96 Add Gtk Version definition mechanism
This allows setting the gtk version specifically
via 1 flags:
`-d:gtkminor=X`.

This sets the constant GtkMinor (default: 0).
There is no GtkMajor version as this is implicitly assumed
to always be 4, as this library is specifically written
for GTK 4.
The default assumed GTK version is thus 4.0.

To compile for a gtk version of 4.8 or higher, simply compile with
`-d:gtkminor=4`.