Disable auto-dismissing toast after timeout
I currently have no solution for the use-after-free
conundrum.
That conundrum is caused by setting a timeout in the
owlkettle layer to dismiss the toast after it's timeout expires.

If you do that and the user dismisses the toast, then the
toast-memory gets free'd after a while by gtk.
If you try to dismiss that toast then after the timeout, you access
memory that was free'd, causing a user-after-free segfault.

I have a feeling gtk should be the one that dismisses these toasts
automatically after the timeout expires,
but I can't get that to work.