Add native approach
Note: This will segfault if you summon a toast,
wait for it to be dismissed
and then try to summon the next one.

The reason for this is that ToastOverlay will free the toast after dismissing it.
If you then swap out the ref that'll trigger the destructor hook,
which will try to unref a pointer to freed memory,
causing a use-after-free error.