Move over toast approach to custom object
That object is only converted to an AdwToast
at the last moment before handing it over
to ToastOverlay.
This way things that belong to gtk are only created when they're
definitely going into to ownership of the gtk library.
Thus no memory leak even conceptually can occur,
as nim-types will automatically be memory managed.