Refactor playground to not require toListFormField
This changes how this all works.
Previously we generated expressions that act directly on `state`
to manipulate its values.

Now we fetch a pointer to each field on state or
each field of an instance in a seq and act on that.
This gets rid entirely of "toFormField", which can now be used
for an individual field as well as a field of an instance in a seq.