Add Variant toFormField overload
This adds a new overload for toFormField to specifically
handle object variants.

Object Variants are hard to handle and thus by default shall receive
the placeholder field that reminds the user to define their own
toFormField proc for their variant.

The Overload works via having a ObjectVariant concept, which catches all
toFormField invocations for object variants.
It simply checks on a technicality if a type-field can be assigned
to itself. This is never true for kind fields on object variants as it
results in compiletime errors.