Refactor playground to not use toListFormField
* Beautify Scale example

* Remove unused import

* Remove unused imports from example

* Refactor playground to not require toListFormField

This changes how this all works.
Previously we generated expressions that act directly on `state`
to manipulate its values.

Now we fetch a pointer to each field on state or
each field of an instance in a seq and act on that.
This gets rid entirely of "toFormField", which can now be used
for an individual field as well as a field of an instance in a seq.

* Fix toFormfield signature for enums

Nim 1.6.12 appears to not like the `ptr enum` type.
ptr[enum] seems okay though, so it just doesn't like the notation.

* Remove requirement for fieldnames to be static in playground

* Add percent formfield

* Add delete button to seq formfield

* Move generics over to use Viewable

This required moving open to widgets.nim as due to being part
of playground that proc was not available there (?).

* Make range toFormField generic over all ranges

This requires introducing a `Range` concept as
`range` is not a valid typeclass to use as a generic parameter.

* Make toFormField doc comments more uniform

* Add index to form field label of sequences

* Add forward declarations for all toFormField procs

This is done in order to avoid any of the following
proc bodies accidentally missing out
on a proc that was provided befor e it.
This helps with not needing to care about the order that the
procs are defined in.

* Add toFormField proc to fully generate form for any type

* Add toFormField proc for any distinct type

* Add back removed import

This import is relevant when -d:pixbufAsync is set

* Add Variant toFormField overload

This adds a new overload for toFormField to specifically
handle object variants.

Object Variants are hard to handle and thus by default shall receive
the placeholder field that reminds the user to define their own
toFormField proc for their variant.

The Overload works via having a ObjectVariant concept, which catches all
toFormField invocations for object variants.
It simply checks on a technicality if a type-field can be assigned
to itself. This is never true for kind fields on object variants as it
results in compiletime errors.

* Change name of ObjectVariant concept

The new name is ObjectVariantType.
This serves as an important distinction that his is a concept that
covers **types**, not **instances**.
The doc comment was adjusted to clarify this as well.

* Fix deletion not working

* Beautify how delete button gets added

* Add support for ref object variants

To do so the proc that checks it just needs to also accept
ref objects and use the `getIterator` template
to decide on whether or not to deref before using fieldPairs.