#115 Add the ability to have circular References
This introduces the StateRef type which is a ref-type containing a reference to a Widget.
It is used for the pseudo field "stateRef" for each Widget, which for any given Widget will contain a reference to said widget.
This allows accessing references within Widgets and passing them back and forth.

It is reactive, to allow subscribing to it and getting updated every time the reference changes.