#115 Improve handling of stateRef changes
Previously the stateRef got the observer unsubscribed and resubscribed every update.
That shouldn't happen, we only want to do that when the stateRef actually change.
Therefore we now check first for changes in the ref.

That uncovered that we need to also do the initial subscription.
This is now handled in the build hook for this field.